--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "editor.tabSize": 2
-}
