+++ app/components/zzzGitLayout/zzzGitLayout.tsx
@@ -1,0 +1,23 @@
+// // import { Metadata } from 'next'
+// import JumbotronCattoFlexible from "@/app/components/JumbotronCattoFlexible/JumbotronCattoFlexible"
+// import SideNavGitCatto from '@/app/components/SideNavGitCatto/SideNavGitCatto'
+
+// const GitLayout = () => {
+//   return (
+//     <>
+//       <div className="flex h-screen">
+//         <aside className="flex-[20]" aria-label="Sidebar">
+//           <SideNavGitCatto />
+//         </aside>
+//         <article className="flex-[80] bg-blue-500">
+//           <JumbotronCattoFlexible
+//             title="Git Source Control"
+//             description="Git is the industry standard for source control. This section I am displaying some of the commands that I use frequently."
+//           />
+//         </article>
+//       </div>
+//       <hr className="m-3" />
+//     </>
+//   )
+// }
+// export default GitLayout;
